Improvement of audio stream management: addition of Audirvana pre-decoded stream detection and adjustment of DSD decoding modes. Change of bufferSeconds parameter type from int to float for better precision. Update of debug messages for increased clarity.

+++ src/DirettaRenderer.h
@@ -18,7 +18,7 @@ public:
         int port;
         std::string uuid;
         bool gaplessEnabled;
-        int bufferSeconds;
+        float bufferSeconds;  // Changed from int to float (v1.0.9)
         int targetIndex;  // -1 = interactive selection, >= 0 = specific target
         
         Config();
